add two numbers

+++ 0x07-python-test_driven_development/0-add_integer.py
@@ -12,4 +12,4 @@ def add_integer(a, b=98):
     elif (type(b) != int and type(b) != float):
         raise TypeError("b must be an integer")
     else:
-        return int((float(a) + float(b)))
+        return (int(a) + int(b))
